Move GEO_IN_GROUP types to pure file
Fixes #20392
