+++ extensions/amp-geo/0.1/amp-geo.js
@@ -48,21 +48,13 @@ import {Services} from '../../../src/services';
  */
 import {ampGeoPresets} from './amp-geo-presets';
 
+import {GEO_IN_GROUP} from './amp-geo-in-group';
 import {dev, userAssert} from '../../../src/log';
 import {getMode} from '../../../src/mode';
 import {isArray, isObject} from '../../../src/types';
 import {isCanary} from '../../../src/experiments';
 import {isJsonScriptTag, waitForBodyPromise} from '../../../src/dom';
 import {tryParseJson} from '../../../src/json';
-
-/**
- * @enum {number}
- */
-export const GEO_IN_GROUP = {
-  NOT_DEFINED: 1,
-  IN: 2,
-  NOT_IN: 3,
-};
 
 /** @const */
 const TAG = 'amp-geo';
